fixed header of homepage, working on modal
